Solid tubes: Make floor thinner and catcher taller
The inner "tube catcher" is too short to hold the slender pipette
tapered ends without letting them escape the bottom.  Increase the
height of the inner tube catcher interface to raise the tubes higher
off the floor.

Note the original sample design has solid floors so the taper did not
get a chance to escape.